requirements.txt
python-dotenv>=1.0.0
websockets>=12.0
edge-tts>=6.1.0
openai-whisper>=20231117
pyaudio>=0.2.14
numpy>=1.26.0
requests>=2.31.0
